Properly support `import <Module> from 'true-myth/<module>'`.
Use TypeScript's [declaration merging][1] functionality to make the
imports behave the way you'd optimally want them to:

    import Maybe from 'true-myth/maybe';
    const foo: Maybe<number> = Maybe.of(12);

Previously, to get this behavior, we had to do some annoying shenanigans
with type assignment, because TypeScript has [a discrepancy][2] in the
ways it treats item name conflicts in imports vs. assignments. To get
the same behavior previously we had to do this:

    import * as Maybe from 'true-myth/maybe';
    type Maybe<T> = Maybe.Maybe<T>;
    const foo: Maybe<number> = Maybe.of(12);

The new style of imports continues to work in a backwards compatible
fashion, but there is no reason to use the `import * as Maybe` variant,
since `import Maybe` provides the same access to all internal members
*and* doubles as the type.

[1]: https://www.typescriptlang.org/docs/handbook/declaration-merging.html
[2]: https://github.com/Microsoft/TypeScript/issues/19553#issuecomment-340216595