Fix type for `Maybe.last`.
Previously, `Maybe.last` would return a `Maybe<T | null | undefined>`
if you ran it on an array whose type was `Array<T | null | undefined>`;
this actually failed to match the run-time behavior (though it was
*looser* than runtime behavior and therefore could not cause a bug in
that direction). It now correctly returns `Maybe<T>` in that scenario.

cf. d9a60e7f, which did the same thing for `Maybe.head`; I should have
remembered to check this then but... didn't.