Support both `Maybe.get` and `Maybe.property`.
Per discussion with @bmakuh and @tansongyang, use `get` for getting a
property off of a `Maybe` instance and `property` for getting a property
off of a regular object. This maintains the symmetry between the
instance method and standalone function version of `get`, and also
provides a convenience tool for non-`Maybe` types (such that `get` is
actually implemented in terms of `property` for both the function and
method variants).