Add `Maybe.wrap` for functions which return `null | undefined`.
Many JS libraries return `null` or `undefined` when they have invalid
values; this convenience function allows a user to wrap those unsafe
functions in a function which works identically but returns a `Maybe<T>`
instead of `T | null | undefined`. For example:

    const querySelector = Maybe.wrap(document.querySelector.bind(this))
    const isVisible = querySelector('#foo').map(el => el.style.display)

The original `Document#querySelector` function returns an `Element` for
the first matching element in the document, or `null` if there are none.
The wrapped function returns `Maybe<Element>` instead.