Switch to expect-type for type assertions
The `assertType` helper we used previously mostly worked for most type
checks most of the time. However, it was weak against (mis)use of `any`
and failed to catch a number of issues that richer assertion libraries
*can* catch. Switching to `expect-type` makes our type assertion story
more robust and thereby sets the groundwork for more substantial rework
required to migrate to newer versions of TypeScript.

Future PRs will further improve the testing story for this library:

- importing from the public API of the library, to guarantee that the
  tests represent the actual end user experience
- using `ember-try` to run tests against a matrix of supported
  TypeScript versions, including `typescript@next` to catch breaking
  changes *before* they happen