Create GH Actions CI Workflow
This config runs the test suite against all supported Node versions on
Ubuntu, and against the version in the `package.json` for macOS and
Windows. (Running against all supported versions for all OSes would
likely take a *very* long time, just because Windows is super slow. It's
worth checking later, though.) The config uses Volta to manage the Node
versions, so that we can be guaranteed it will use the same toolling as
development.

Resolves #61