Introduce `true-myth/toolbelt` module for better tree shaking
Extract the helpers which couple `Maybe` and `Result` to each other
into the new module, and re-implement the existing implementations as
re-exports of the new canonical implementations.

This *very slightly* decreases the convenience of swapping between the
two, but makes it possible to use either without reference to the
other. Particularly in a world with optional chaining, nullish
coalescing, and nullish assignment, many end users of the library may
find more value in `Result` than in `Maybe`, and until now it has been
impossible to use one without paying for the other, because the classes
themselves have referenced the other module.

In a future change, we can deprecate the legacy versions. Once users
migrate their existing uses to using the new `toolbelt` module, we can
safely remove them in 6.0 (next year), at which point users who only
use `Maybe` or `Result` will be able to ship bundles which do not
include the other at all.