Remove ESLint
ESLint is a great tool. Also we don't need it. This is a project
maintained by two very experienced engineers with, between the two of
us, nearly a decade of experience writing TypeScript and about fifteen
years of writing front-end code. It's not that linting isn't valuable
in many context. It's that it isn't valuable in *this* context.