Fix regression of `Result.ok()`
In the work done to collapse the separate classes together for 5.0, I
accidentally kept the *wrong* side of an inconsistency from all
versions prior to v5. Historically we had a bit of a weird misalignment
around the constructors: you could call `Result.ok()`, which was
actually a sneakily-exported standalone function which constructed an
`Ok` class instance, or you could call `new Ok()`. If you called
`Result.ok()`, you got back `Result<Unit, unknown>`. If you called `new
Ok()`, you got the error you hit here. When I folded the `Ok` and `Err`
classes into a single `Result` class, I accidentally brought along the
error instead of the convenience function behavior!

While making the fix here, introduce additional explicitness so that
the runtime behavior and the type behavior match: calling the `ok` or
`err` static constructors with `null` or `undefined` produces a
`Result` with a `Unit` value in that position instead of ending up with
`Result<null, E>`. (This requires a single unsafe cast, in the case of
the `toolbelt` function `transposeMaybe`.)

Resolves #261