(Forward port) Fix v5 regression of checks on `.value`
Here, the change between v5 and v6 of how we handle the internal
representation of the types enabled us to use direct comparison between
the two `Maybe` values in the comparison, but the use of `Pick` and
`Exclude` utility types (to fix #268) excludes private types, which in
turn makes `repr` invisible when using `Maybe<T>` as the type of the
comparison. Here, simply cast *without* doing any runtime operations so
that the check can remain as it is.