(Forward port) Revert handling of explicit `null` and `undefined`
While the new behavior is preferable, because it means you never end up
with `Result<null, E>` or `Result<undefined, E>` etc., it does not
actually work correctly with inference: the type parameter resolution
consistently ends up bailing and ending up as `Maybe<unknown>` when it
should be resolving to `Maybe<T>` within a `Result<Maybe<_>, E>` (or
vice versa in the `Err` case).

While it is probably worth thinking about how to tackle this again in
v6, it may just be a persistent limitation of the overload-based type
resolution that this requires.

Fixes #268