Incorporate toString handling from v5
(cherry picked from commit 25adb1aa0b4c02c7bf980c9db3fac7ab416de7b7)

---

Original message:

Add support for TypeScript 4.7

TS 4.7 catches a class of error which 4.6 and earlier did not, courtesy
of microsoft/TypeScript#43183: if the type passed does not implement
`toString()`, it notices, because it no longer defaults to falling back
to `{}`.

To resolve this, *loosen* the constraints on what is acceptable as
input to `toString`, by making the implementation itself more robust.
Doing so also fixes a bug with the output for `toString` when working
with strings: previously if you had `Maybe("a string")`, the `toString`
output would be `'Maybe(a string)'`; it is now `'Maybe("a string")'`.

Fixes #330