Correct the v6 implementation of `equals`
Because of the use of `Omit` as introduced in b4071ba for [this PR][1]
(backported as 92a1b08), TS can no longer see that the type of `Result`
is the same as `ResultImpl`. Notably, this is the same as the fix made
in 9b777c8 for the same issue with `Maybe`.

[1]: https://github.com/true-myth/true-myth/pull/336