Update publication mechanics with "exports"
Simplify and otherwise update the `tsconfig`s for compilation and
publication strategies:

- Use the new `exports` field in `package.json` to define the public
  API for both `.js` and `.d.ts` files, so that we can entirely drop
  the special handling for generating types in the root.

- Define a single shared `base.tsconfig.json` file from which all
  others inherit, and update the `tsconfig.json`

- While updating the base config, add `noUncheckedIndexedAccess` and
  `exactOptionalPropertyTypes` to bring the package into maximal
  strictness (and conformance with <https://www.semver-ts.org>).

- Update the Jest and TypeDoc configs to take advantage of the new
  configuration, while maintaining the same actual public API.

- Explicitly drop support for TS <4.7 on this branch, but maintain it
  on the v5.x branch.

Note: this does *not* change the public API at all. It should "just
work"... for anyone who is using Node 14 and TS 4.7.

In other words: it is a breaking change, but one which consists
entirely of people needing to update their TS and Node versions. This
means this cannot go out until TS 4.7 is out and stable and people can
separately upgrade to a True Myth v5.x release targeting TS 4.7.