Decrease Unit implementation runtime size
Make it a prototype-less object which is *cast* as `Unit`. This means
that it has an even smaller overhead at runtime than it did before, and
it basically acts like an empty object in the type system... but one
with a useful name, and a guarantee that there is only ever one of it.