Add type-safe `Ok.cast()` and `Err.cast()` methods
When working with `Result`, users sometimes find it useful to cast the
*other* case of the type. For example, given an `Ok<string, number>`
(note: not a `Result<string, number>`), casting to `Ok<string, string>`
is a safe no-op cast: we *know* we are in the `Ok` case, not the `Err`
case, so the transformation is guaranteed to be legal. This is exactly
equivalent to two other potential type representations, in fact:

- `Ok<T> | Err<E>`
- `Ok<T, never> | Err<never, E>`

Implementing either of those types with the performance guarantees we
currently have for True Myth *and* preserving type safety throughout
is unfortunately not straightforward (or necessarily even possible),
but `.cast()` allows us to get the benefits anyway.