Publish True Myth as a dual-mode package (ESM + CJS)
Add a new `tsconfig` for the CJS exports, and update the `exports` map
in `package.json` to cover both ESM and CJS. Then introduce a small
publishing script which can manage this for us, which:

- Cleans up `dist`.
- Runs `tsc` against the two configs.
- Renames all the files in the CJS export. 🤦🏻‍♂️

Along the way, restructure the package contents to exclude access to
`-private` *without* having an intermediate `public` directory (using
the `"path": null` approach), which will help TypeDoc emit the right
thing, and make sure that everything is accessible from the index by
re-exporting the contents as namespaces. (We will likely want to
evaluate this specific mechanic before publishing a version which
includes this! In particular: the names?)

Note: this implementation does fall afoul of *one* of the hazards of
dual-mode packages: there are simply two different published bundles on
disk, which means that the actual classes from ESM and CJS will differ.
In most cases, this does not matter at all, since the only symptom of
that is that `instanceof` checks will not match. Given the way this
library is expected to be used (i.e. totally statelessly and for pure
functional programming), this is deemed an acceptable tradeoff, since
otherwise the level of shenanigans involved is very high.

Publish True Myth as a dual-mode package (ESM + CJS)

Add a new `tsconfig` for the CJS exports, and update the `exports` map
in `package.json` to cover both ESM and CJS. Then introduce a small
publishing script which can manage this for us, which:

- Cleans up `dist`.
- Runs `tsc` against the two configs.
- Renames all the files in the CJS export. 🤦🏻‍♂️

Along the way, restructure the package contents to exclude access to
`-private` *without* having an intermediate `public` directory (using
the `"path": null` approach), which will help TypeDoc emit the right
thing, and make sure that everything is accessible from the index by
re-exporting the contents as namespaces. (We will likely want to
evaluate this specific mechanic before publishing a version which
includes this! In particular: the names?)

Note: this implementation does fall afoul of *one* of the hazards of
dual-mode packages: there are simply two different published bundles on
disk, which means that the actual classes from ESM and CJS will differ.
In most cases, this does not matter at all, since the only symptom of
that is that `instanceof` checks will not match. Given the way this
library is expected to be used (i.e. totally statelessly and for pure
functional programming), this is deemed an acceptable tradeoff, since
otherwise the level of shenanigans involved is very high.