Drop explicit `.d.ts(.map)` from `files`
As is, this was misleadig; it is a holdover from loooong ago when we
used Ember CLI to publish and needed to include those files explicitly,
rather than just pointing to `dist`. It *implied* that we were *not*
including the new `.d.cts` etc. files, or for that matter `.js` files;
but we were, since they are included in `dist`.