Handle `undefined` correctly with `Result.ok`
Previously, we treated the arity-0 and the arity-1 with a first argument
of `undefined` versions of `Result.ok` identically. This can directly
result in a type error at runtime by setting the actual value to `Unit`
while “successfully” producing a type of `Result<undefined, SomeError>`,
since there was a mismatch between the implementation and the produced
type (i.e. a case where the overload logic and the implementation logic
diverged):

```ts
function foo(): Result<undefined, string> {
    return Result.ok(undefined);
}
```

Since the implementation here simply checks whether the argument
`isVoid`, it will use `Unit` under the hood; but the type checking
correctly sees that we are passing a `T` which is `undefined` and thus
produces the correct type. The result is that the `.value` is `Unit`
when it *should* be `undefined`, which is the inverse of the case we had
previously taken great care to handle.

This solves the problem by changing the behavior so the public contract
and the implementation actually match, by doing an arity check instead
of an `isVoid` check on the first argument. As a result:

- `Result.ok()` → `Result<Unit, E>`
- `Result.ok(undefined)` → `Result<undefined, E>`

Fixes #607