Prevent `Maybe.map` and `Maybe.ap` from returning `null`
In the case of `Maybe.map`, this is straightforward, by applying a type
constraint on the `U` type paramater. With `Maybe.ap`, the same holds,
but it is also important to prevent users from even working with a
lifted function type which will produce this, so also further introduce
dedicated overrides for passing functions to `Maybe.of` or `Maybe.just`,
which produce `never` rather than a `Maybe<() => null>`, so that it is
impossible to use those types at all.

It is, unfortunately, not possible to prevent the construction of those
types directly, since there is no way to constrain a generic to allow
“anything but a function”.