Forbid `null` and `undefined` as arguments to `Maybe.just`
While working on #753, I noticed that a similar issue applies to the
`Maybe.just` constructor as to `Maybe.of`: it historically would allow
functions which produce `null` or `undefined`. I included the fix for
that specific issue in #755. However, as I did so, I realized that the
types for `Maybe.just` have been looser than they need to be for quite a
while now: when we wrote the library, we could not properly forbid
`null` or `undefined` at construction time, because we had no good way
to name the constraint that any argument passed to `just` be non-null.
That changed several years ago, so we can now require only types which
are `{}`, i.e. not `null` or `undefined`, on the type side while also
continuing to throw if someone passes incorrectly typed data which is
actually `null` or `undefined`.

This is not a breaking change, but a *feature*: it is what we always
wanted, and aligns the types with the runtime behavior. However, it may
in some cases require additional type annotations when working directly
with a `Result` and using toolbelt functions. Specifically, there were
some cases where the old implementation unsafely allowed `unknown`,
which necessarily includes `null | undefined`, and could therefore end
up with a `Maybe<unknown>`, and indeed via the unconstrained generic
could move from `Result<null, string>` to `Maybe<null>` using e.g. the
`transposeResult` method. This is no longer possible! 🎉