Update documentation comments for Typedoc 0.26
This currently optimizes for folks reading the Typedoc-generated HTML
versions of the docs, though the ways it does show primarily show up in
the `toolbelt` module: cross-module linking is still a thing Typedoc
seems to struggle a bit with in the case that you want to reference a
term which is not in scope. I may be able to work around that, but for
now this is the best I can do.