feature: add `unwrap` and `unwrapErr` test helpers
Long ago, we had methods which did this:

- `Maybe.prototype.unsafelyUnwrap`
- `Result.prototype.unsafelyUnwrap`
- `Result.prototype.unsafelyUnwrapErr`

We removed these when we made it possible to directly access `.value`
and (for `Result`) `.error` after narrowing, because that is always safe
to do and guides users toward a happy path. However, it can be rather
annoying for tests, not least for tests where “does this return the
correct variant” is the thing under test!

Because we want to continue to encourage the safe path as the happy path
and to avoid adding any bloat to the payload for users who do *not* use
these, add `unwrap` and `unwrapErr` functions to a new `test-support`
module. The `unwrap` function is polymorphic; it works correctly with
both `Maybe` and `Result`. For obvious reasons, `unwrapErr` only applies
to a `Result`.

Fixes #817