Result: correct the implementation of `static err` constructor
Match the `Result.ok` constructor to correctly (…) allow `null` or
`undefined` to be passed as an explicit `Err` value. This is not
something we *encourage*, but it is supported.

Note: The `never` term here was never (no pun intended!) going to cause
an actual *problem*, because of how the rest of the inference chain was
working, but it was also a strange default. Using `unknown` is better.