Task: handle errors thrown in executor in constructor
We do not go out of our way to add special error handling in general,
taking more of a stance that we trust the people using this understand
the whole point is to use these tools instead of throwing errors or
using `null` or `undefined`. In the case of the executor passed to the
`Task` constructor, though, it is nice to provide some more useful and
actionable feedback, since it differs significantly from the behavior of
the `Promise` constructor, although it *looks* the same. The `Promise`
constructor absorbs errors thrown during the executor into a rejection,
but `Task` cannot do that (at least, not without mutilating its API).

Instead, catch any such error, wrap it in `TaskExecutorException`, an
`Error` subclass` dedicated to this which explains what went wrong and
where it went wrong and what the solution is, and re-throw it.