Task: use the `const` + `type` export hack
This approach, which we have used for `Maybe` and `Result` for a few
years now, has two benefits:

1. Most importantly, it allows the *type* to be defined as a union:

    ```ts
    type Task<T, E> = Pending<T, E> | Resolved<T, E> | Rejected<T, E>`;
    ```
    
    This in turn allows those types to work well with type narrowing, by
    making those types include, or not include, the `value` and `reason`
    properties at a type level, and requires the user to check the state
    of the `Task` before they can directly access those values.
    
2. The exported item is not sub-classable (from a TS point of view at
   least). This matches our intent. While this is not a primary goal, it
   is a nice side effect of the choice made for (1).

The downside is that it makes the docs less elegant. I deem the upside
from a safety point-of-view to be worth that tradeoff, not least because
we can do other things to help with the documentation.