Task: add top-level `fromResult`, `fromPromise`, and `fromUnsafePromise`
I am adding these (and indeed keeping them) because the reality is that
we cannot handle *all* error possibilities within this library. Although
having (and preferring!) the `tryOr` and `tryOrElse` utilities is good,
the reality is that the same motivations for those applies to every
function in the library. `Maybe.map` will have problems if the callback
passed to it throws an error; so will `Result.ok`. Having these less-
safe helpers is still good on the whole. Adding documentation for them
which makes explicit this tradeoff and refers to the safe alternatives
for the cases where it matters is pretty much the best we can do.