Task: module-scoped versions of instance methods
Introducing these brings the `Task` module to full parity with the
`Maybe` and `Result` modules. As the documentation notes, however, these
are not great to use in practice because they require explicitly setting
the types at the call site, like so

```ts
import * as Task from 'true-myth/task';
let toLength = Task.map<string, number, Error>((s) => s.length);
```

This is annoying, but I think it is worth having for the sake of parity.
I suspect we may want to remove all auto-currying at some point in the
not-too-distant future, because it just does not work very well given
the limits of TypeScript’s type inference. Alas!