Task: add a new `withRetries` function
This new utility function, modeled fairly directly on the [the Rust
`retry` crate][crate], lets you retry a `Task` if it initially rejects
by supplying a `Strategy` which is simply an iterator which produces a
number of milliseconds to wait to try again. When combined with iterator
helpers (either manually authored via generator functions or via the new
iterator built-ins in ES2025), this makes for a very straightforward way
to build up retry logic.

```ts
import * as Task from 'true-myth/task';
import * as Delay from 'true-myth/task/delay';

let fetchTask = Task.withrRetries(
  () => Task.fromPromise(fetch('https://true-myth.js.org')),
  Delay.exponential(10, { factor: 3 }).map(Delay.jitter).take(10),
);
```

(This same strategy works even in versions of TS and runtimes prior to
the stabilization of the Iterator Helpers proposal in ES2025, using ES6
generators for `take`, `map`, etc.)

[crate]: https://docs.rs/retry/latest/retry/