Bug: `all` and `any` should preserve order of types
When I originally implemented these, I was rusty on the relevant type-
level shenanigans that would be required to make them type check, but an
unrelated bit of TypeScript work reminded me that we can use variadic
generics in exactly this kind of scenario: we “just” need to constrain
the relevant type aliases to use `[...TheRelevantType]` instead of using
`[number]` to index them as arrays, and then also update the
implementations to preserve that ordering.

The runtime implication is small:

1. Use `Array.prototype.entries` to enumerate the tasks and thereby save
   which `Task` is resolving or rejecting at any given point in `all` or
   `any` respectively.

2. Pre-allocate an array with the same length of the original array of
   tasks, so that it can be filled in at the appropriate location. This
   implies that the array will be “sparse” during the lifecycle of the
   function unless or until all the tasks resolve or reject for `all`
   and `any` respectively. This will never be visible to the caller, as
   the array will be returned full *or* garbage collected as unused.

In most cases, I expect the potential additional allocation overhead to
be irrelevant: if anything, we may end up allocating *smaller* arrays
this way

This is a bug fix that *will* produce red squiggles, but of the sort the
SemVer TS spec expressly allows: they represent code users can *delete*,
because they *had* to do runtime checks before to figure out what the
type was for each item in the array, i.e., where they previously might
have ended up with a `Resolved<Array<string | number>>`, they now might
end up with `Resolved<[string, number]>`, so `task.value[0]` will be
statically known to be `string` rather than `string | number`.

I started by fixing it here as targeting the v8.x branch, because it is
unexpected and does not match the types for `Promise`; I will also port
it forward to the `main` branch targeting the upcoming v9.0 release.