Bug: `safe` should require `PromiseLike` return types
By allowing non-`PromiseLike` types, this caused an internal error in
True Myth’s own implementation. This is a significant downside of using
overloads, as well as the unsafe cast of the return type that previously
existed in the body of `safe`. Here, get rid of that cast by requiring
the argument to be a function that returns a `PromiseLike<T>`, loosen
the requirement on `safelyTryOrElse` to require only a `PromiseLike<T>`
rather than a `Promise<T>` (given that is all its implementation needs),
and add type tests to confirm that this  now produces the expected type
errors for functions that do not produce a `PromiseLike`.

Fixes #961