Maybe: correctly handle arrays which include null and undefined
With the `first` and `last` helpers, we historically, wrongly, treated
these two cases as both producing `Nothing`:

```ts
import { last } from 'true-myth/maybe';

let empty = last([]);
let hasNull = last([null]);
```

These are *not* the same, though! One of them *has* an item in it that
is `null`; the other is *empty*.

Update the definitions of `first` and `last` so they now produce nested
`Maybe<Maybe<T>>`. Although this is slightly more annoying, it is more
correct, and can be flattened easily with `andThen`:

```ts
import { first } from 'true-myth/maybe';

let empty = first([]).andThen((v) => v); // Nothing
let hasUndef = first([undefined]).andThen((v) => v); Nothing
```

Note: This is a bug fix, but it is also *quite* breaking. As such, it
will be part of 9.0 and *not* backported to the 8.x branch.