Task: define `Strategy` as `Iterator<number>`
This appears to *actually* work as my previous change was *intended* to
work. The trick is updating the implementation to only use the API of
`Iterator<T>` instead of `IterableIterator<T>`. (This collection of APIs
is very weird, among my least favorite in JS, and it is not particularly
clear in any documentation: MDN, TypeScript, or otherwise.) All custom
strategies that used the original `IterableIterator`-based definition
(i.e., any code written against the API up until 8.5.1’s mistaken “fix”)
will still work with this change.