Feature: generalize inference for `andThen` and `orElse`
Introduce some type machinery to support inferring a distributive type
in the output of `andThen` or `orElse` for all the core types, while
avoiding circularity in the type definitions.

The fundamental goal here is to make it so that when a user returns a
type that is technically a union of the class, `Maybe<A> | Maybe<B>`, we
instead produce a union of the type parameter, `Maybe<A | B>`. This can
come up quite easily when a user is building up a safe abstraction in
idiomatic TS. For example (in line with the user issue that showed us
the need for this change), you might be building up a pre-configured web
“client” that simply returns object types inline:

```ts
import * as task from 'true-myth/result';

let fetch = safe(window.fetch, (cause) => ({
  kind: 'NetworkError' as const,
  cause,
}));

let get = (url: string) =>
  fetch(new URL(url)).andThen((response) => {
    if (response.status >= 400 && response.status < 500) {
      return Task.reject({
        kind: 'ClientError' as const,
        response,
      });
    }

    if (response.status >= 500 && response.status < 600) {
      return Task.reject({
        kind: 'ServerError' as const,
        response,
      });
    }

    return Task.resolve(response);
  });

let theResult = await get('https://true-myth.js.org');
```

Prior to this change, the type collapsed to `Task<unknown, unknown>`
unless constrained explicitly at the `andThen` call site, because the
type returned here as TypeScript sees it is not a single `Task` with a
union of the custom error object types, but rather multiple distinct
`Task` types: one for each custom error object type! This is technically
*accurate*, but in practice is not what people expect, and because we
constrain the return type to be a `Task`, it is always safe in practice
for people to treat it as `Task<A | B>` instead of `Task<A> | Task<B>`,
and likewise for the equivalent `Maybe` and `Result` types.

We can get the desired behavior with conditional distributive types. By
using a constrained type parameter for the whole returned wrapper type,
rather than the fixed type parameters for the wrapper type, we can
distribute unions back into the type. For example, with `Task`:

```ts
class Task<T, E> {
  // ...
  
  andThen<R extends Task<unknown, unknown>>(
    fn: (t: T) => R
  ): R extends Task<infer U, infer F> ? Task<U, E | F> : never;
}
```

This type will always safely and correctly produce a `Task<U, E | F>`
because of the initial constraint, and indeed will allow gathering up
additional types as it goes along with the original error type, but both
the `U` and `F` types produced here will be distributed over any union
(explicit or implicit) in the return type of the passed `fn`.

Unfortunately, that code as written *does not work* because we define
the types *not* as the classes but instead as the union of types derived
from them. With `Maybe`, for example:

```ts
interface Just<T> extends MaybeImpl<T> {
  readonly variant: 'Just';
  value: T;
  isJust: true;
  isNothing: false;
}

interface Nothing<T> extends Omit<MaybeImpl<T>, 'value'> {
  readonly variant: 'Nothing';
  isJust: false;
  isNothing: true;
}

type Maybe<T> = Just<T> | Nothing<T>;
```

If we now try to define the return type of `MaybeImpl.prototype.andThen`
with this definition of `Maybe`, TypeScript will complain about circular
definition of the `Maybe` type, and likewise for `Result` and `Task`.

To work around that, introduce a small bit of additional type machinery:

- Type-only (i.e. `declare`-only) symbols, `IsMaybe`, `IsResult`, and
  `IsTask`, that are attached to the `Maybe`, `Result`, and `Task`
  classes respectively, using `declare` syntax, and attaching the type
  parameters for the classes to the .

- Corresponding `SomeMaybe`, `SomeResult`, and `SomeTask` types that use
  the declared symbols with their associated use of the type parameters
  to carry all of the relevant type information *without* any additional
  runtime overhead *and* without introducing circularity in the type
  definitions.
  
- A type utility to pluck the type parameters out of those definitions,
  e.g. `TypesFor<S extends AnyTask>`. This provides enough information
  to get the required output for `andThen` and `orElse` without any
  circularity in the definition.

- In `task`, additionally redefine the existing `ResolvesTo` and
  `RejectsWith` utility types in terms of the new utility. Note that
  this is not breaking because those types, as well as the types they
  depend on, are expressly `@internal`.

Fixes #994