build: Fix documentation build, and always check it.
Running `pnpm docs` opens the documentation page for the project, which
is on the one hand maybe a nice shortcut but on the other quite annoying
in this actual case. Use `pnpm run docs` instead.

Additionally, add a CI check to make sure that docs build successfully.
Annoyingly, the dev built for VitePress does *not* do everything that a
prod build does, e.g. link checking: reasonable, but annoying.