Standard Schema: split 3rd-party integration test into own suite
This allows us to test these third-party integrations against the latest
TS, which they should (generally!) work with, independent of our own TS
support matrix. We will need to make a point to document this: we are
explicitly *not* promising that those libraries always work correctly,
only using these tests as a “smoke test” that our implementation of the
Standard Schema works as expected with some of the most popular
libraries that implement conformity with it.