Maybe: constrain wrapped type to be `{}` at construction
Previously, `Maybe` relied on return types to constrain the wrapped type
to be non-nullable (literally by using `NonNullable` on the input type).
Since TypeScript 4.8’s changes/tweaks to assignability with `{}`, it has
been possible to constrain a type to “any non-nullable type” by using
`extends {}`. We had long since enforced safety via `NonNullable`, so we
did not rewrite this along those lines at the time, but it makes for a
cleaner and simpler API, and forbids even writing nonsensical types like
`Maybe<null>` or `Maybe<undefined>`—types that were previously nameable
but not constructible.

Note that this is *not* a breaking change, but it may change where in a
given rejected type the error diagnostic appears.